Moved the axisTable (and its repopulation) to InputConfiguration.
This is so that new InputConfigurations can be created and set up properly for use on the fly.
InputConfigurations and InputManager share the same reference to the dictionary, so inputConfigs can update independently and the manager will see the changes.

InputConfiguration.Duplicate has also been updated.
CreateInputConfiguration needs an alternative method RegisterInputConfiguration for configurations created by means of "new" or "InputConfiguration.Duplicate"